+++ src/manus-web/index.ts
@@ -1,0 +1,1 @@
+export { createManusWebExecutor } from "./executor.js";